rename component Switch to ToggleSwitch
