+++ assets/react/components/Switch/ToggleSwitch.js
